docker/fastapi-backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
joblib
scikit-learn
nltk
google-cloud-secret-manager
google-cloud-storage